edit gallery and styles

+++ src/components/SingleGallery.vue
@@ -3,23 +3,32 @@
 
         <div class="gallery-data">
             <h1>{{ gallery.title }}</h1>
-            <div v-if="gallery.user" style="margin-top: 1rem;">
-                <button 
-                    v-if="isAuthenticated && gallery.user.id == user_id"
-                    @click="deleteGallery(gallery)" 
-                    name="submit" class="btn btn-primary" type="submit">Delete
-                </button> 
+            <div v-if="gallery.user">
+                <div id="buttons">
+                    <button 
+                        v-if="isAuthenticated && gallery.user.id == user_id"
+                        @click="deleteGallery(gallery)" 
+                        name="submit" class="btn btn-danger" type="submit">
+                        Delete
+                    </button>
+                    <router-link 
+                        v-if="isAuthenticated && gallery.user.id == user_id"
+                        :to="{ name: 'edit-gallery', params: {id: gallery.id}}" 
+                        name="submit" class="btn btn-dark" type="submit" id="edit-btn">
+                        Edit
+                    </router-link> 
+                </div>
+                
                 <em>Author:</em> 
                 <router-link :to="{ name: 'single-author', params: {id: gallery.user.id}}" id="user">
                     {{ gallery.user.first_name }} {{ gallery.user.last_name }}                
                 </router-link>
             </div>
             <div><em>Created at:</em> {{ gallery.created_at }}</div>
-            <div style="margin-top: 1rem;">{{ gallery.description }}</div>
+            <div id="description">{{ gallery.description }}</div>
         </div>
 
         <b-carousel id="carousel1"
-            style="text-shadow: 1px 1px 2px #333; margin: 0 auto; margin-top: 2rem; margin-bottom: 3rem;"
             controls
             indicators
             background="#ababab"
@@ -28,29 +37,32 @@
             @sliding-start="onSlideStart"
             @sliding-end="onSlideEnd"
             v-if="gallery.images">
-            <b-carousel-slide id="image"
+            <b-carousel-slide
                 style="height: 500px;"
-                v-for="image in gallery.images" :key="image.id"
-                :img-src="image.image_url"
-            ><button id="link-view" class="btn btn-outline-light" @click="openInNewTab(image.image_url)">View</button> </b-carousel-slide>
+                v-for="image in gallery.images.slice().reverse()" :key="image.id"
+                :img-src="image.image_url">
+            <div id="link-view">
+                <button class="btn btn-dark" @click="openInNewTab(image.image_url)">View</button>
+            </div>
+            </b-carousel-slide>
         </b-carousel>
         
-        <h3 v-if="gallery.comments">Comments:</h3>
-
         <ul class="list-group list-group-flush" v-for="comment in gallery.comments" :key="comment.id">
-            <li class="list-group-item" v-if="comment.user"><strong>
-                {{ comment.user.first_name }} {{ comment.user.last_name }}
-            </strong></li>
-            <li class="list-group-item" style="font-size: 0.8rem;"><em>{{ comment.created_at }}</em></li>
-            <li class="list-group-item">{{ comment.text }}</li> 
-            <div v-if="comment.user">
-                <button 
-                    v-if="isAuthenticated && comment.user.id == user_id"
-                    @click="deleteComment(comment)" 
-                    name="submit" class="btn btn-primary" type="submit">Delete
-                </button> 
-            </div>
-                                             
+            <li class="list-group-item">
+                <div id="first-li" v-if="comment.user"><strong>
+                    {{ comment.user.first_name }} {{ comment.user.last_name }}</strong>
+                    <div v-if="comment.user">
+                        <button 
+                            v-if="isAuthenticated && comment.user.id == user_id"
+                            @click="deleteComment(comment)" 
+                            name="submit" class="btn btn-danger">
+                            <i class="far fa-trash-alt"></i>
+                        </button> 
+                    </div>
+                </div>
+                <div style="font-size: 0.8rem;">at <em>{{ comment.created_at }}</em></div>
+                <div>{{ comment.text }}</div> 
+            </li>                                 
         </ul>
         <form v-if="isAuthenticated" @submit.prevent="onSubmit">
             <div>
@@ -59,7 +71,7 @@
              <div class="input-group">
                 <span class="alert alert-warning" v-if="e.text">{{ e.text[0] }}</span>                        
             </div>
-            <button name="submit" class="btn btn-primary" type="submit">Add comment</button>
+            <button name="submit" class="btn btn-info" type="submit">Add</button>
         </form>
     </div>
 </template>
@@ -77,20 +89,19 @@ export default {
             sliding: null,
             newComment: {},
             e: [],
+            er: '',
             user_id: authService.getUserId(),
             isAuthenticated: authService.isAuthenticated()
         }
     },
 
-    beforeRouteEnter(to, from, next){
-        galleries.get(to.params.id)
+    created(){
+        galleries.get(this.$route.params.id)
         .then(response => {
-            next(vm => {
-                vm.gallery = response.data
-                console.log(response.data)
-            })
+            this.gallery = response.data
+            console.log(this.gallery)
         })
-        .catch(err => console.log(err))
+        
     },
 
     methods: {
@@ -106,6 +117,7 @@ export default {
         },
 
         onSubmit(){
+            
             this.newComment.gallery_id = this.gallery.id
             galleries.addComment(this.newComment)
             .then(response => {
@@ -119,6 +131,9 @@ export default {
         },
 
         deleteComment(comment){
+            if (!confirm('Are you sure you want to delete this comment?')){
+                return;
+            }
             galleries.deleteComment(comment.id)
             .then(response => {
                 let index = this.gallery.comments.indexOf(comment)
@@ -128,6 +143,9 @@ export default {
         },
 
         deleteGallery(gallery){
+            if (!confirm('Are you sure you want to delete this gallery?')){
+                return;
+            }
             galleries.deleteGallery(gallery.id)
             .then(response => {
                 this.$router.push('/my-galleries')
@@ -140,23 +158,66 @@ export default {
 }
 </script>
 
-<style>
+<style scoped>
+
+.container {
+    margin-top: 2rem;
+}
+h1 {
+    margin-bottom: 2rem;
+}
 .gallery-data {
     text-align: center;
 }
 
+#description {
+    margin: 2rem;
+}
+
+#buttons {
+    margin: 0 auto;
+    width: 15%;
+    margin-bottom: 1rem;
+    display: flex;
+    justify-content: space-between;
+}
+
+#edit-btn {
+    width: 4rem;
+}
+
 #user {
-    color: grey;
+    color: black;
+}
+
+#carousel1 {
+    text-shadow: 1px 1px 2px #333; 
+    height: 500px; 
+    margin: 0 auto; 
+    margin-top: 2rem; 
+    margin-bottom: 3rem;
 }
 
 #link-view {
-    /* margin-bottom: 35rem;
-    margin-left: 22rem;  */
+    margin: 0 auto;
+    width: 10%;
+    margin-bottom: 25rem;
+}
+
+.list-group {
+    width: 50%;
+}
+
+#first-li {
+    display: flex;
+    justify-content: space-between;
 }
 
 form{
     margin-top: 2rem;
+    padding-bottom: 3rem;
 }
+
 
 </style>
